// frontend/src/client/types.gen.ts
export type AssetPresignedUploadRequest = {
    /**
     * File Name
     */
    file_name: string;
    /**
     * Content Type
     */
    content_type: string;
    /**
     * Size
     */
    size: number;
};

/**
 * AssetPresignedUploadResponse
 */
export type AssetPresignedUploadResponse = {
    asset: AssetPublic;
    /**
     * Upload Url
     */
    upload_url: string;
    /**
     * Method
     */
    method?: string;
    /**
     * Expires In
     */
    expires_in: number;
    /**
     * Required Headers
     */
    required_headers: {
        [key: string]: string;
    };
};

/**
 * AssetPublic
 */
export type AssetPublic = {
    /**
     * File Name
     */
    file_name: string;
    /**
     * Content Type
     */
    content_type: string;
    /**
     * Size
     */
    size: number;
    /**
     * Id
     */
    id: string;
    /**
     * Owner Id
     */
    owner_id: string;
    /**
     * Object Key
     */
    object_key: string;
    /**
     * Created At
     */
    created_at: string;
};